+++ tool/src/ase.ts
@@ -16,6 +16,7 @@ import DiagramCommand              from "./ase-diagram.js"
 import SetupCommand                from "./ase-setup.js"
 import StatuslineCommand           from "./ase-statusline.js"
 import TaskCommand                 from "./ase-task.js"
+import ArtifactCommand             from "./ase-artifact.js"
 import pkg                         from "../package.json" with { type: "json" }
 
 /*  type of top-level (global) options  */
@@ -63,6 +64,7 @@ const main = async (): Promise<void> => {
     new HookCommand(log).register(program)
     new StatuslineCommand(log).register(program)
     new TaskCommand(log).register(program)
+    new ArtifactCommand(log).register(program)
     new DiagramCommand(log).register(program)
 
     /*  parse program arguments  */